create favourite-manager module
